requirements.txt
# ================================
# Hacker News 爬虫系统依赖包
# ================================

# 核心依赖 (Core Dependencies)
# ================================

# 网络请求库 - 用于爬取网页和API调用
requests>=2.31.0

# HTML解析库 - 用于解析HN网页内容
beautifulsoup4>=4.12.0

# 数据处理库 - 用于CSV数据操作
pandas>=2.0.0

# 环境变量管理 - 用于配置文件加载
python-dotenv>=1.0.0

# Telegram Bot库 - 用于消息推送
python-telegram-bot>=20.0

# 异步HTTP客户端 - 用于异步网络请求
httpx>=0.24.0

# 任务调度库 - 用于定时任务
schedule>=1.2.0

# 开发依赖 (Development Dependencies)
# ================================

# 代码格式化工具
# black>=23.0.0

# 代码检查工具
# flake8>=6.0.0

# 类型检查工具
# mypy>=1.0.0

# 测试框架
# pytest>=7.0.0

# 可选依赖 (Optional Dependencies)
# ================================

# 更好的JSON处理
# ujson>=5.7.0

# 更快的CSV处理
# pyarrow>=12.0.0

# 内存监控
# psutil>=5.9.0

# 配置验证
# pydantic>=2.0.0

# 日志增强
# loguru>=0.7.0

# ================================
# 安装说明
# ================================
# 
# 基础安装：
# pip install -r requirements.txt
# 
# 开发环境安装（包含开发工具）：
# pip install -r requirements.txt black flake8 mypy pytest
# 
# 性能优化安装（包含可选依赖）：
# pip install -r requirements.txt ujson pyarrow psutil
# 
# 完整安装：
# pip install -r requirements.txt black flake8 mypy pytest ujson pyarrow psutil pydantic loguru
# 
# ================================
# 版本兼容性
# ================================
# 
# Python: 3.8+
# 操作系统: Linux, macOS, Windows
# 
# 注意事项：
# 1. 某些依赖可能需要编译，确保系统有相应的编译工具
# 2. 在某些系统上可能需要额外安装系统依赖
# 3. 建议使用虚拟环境安装依赖
